pyproject.toml
[tool.poetry]
name = "diplomabackend"
version = "0.1.0"
description = ""
authors = ["Dmitrii <stakheev12@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
path = "^16.9.0"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"

requests = "^2.27.1"
python-dotenv = "^1.0.0"

loguru = "^0.7.2"
aiohttp = "^3.9.3"
asyncio = "^3.4.3"

pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"
gunicorn = "^21.2.0"
strawberry-graphql = "^0.217.0"
graphql-core = "^3.2.3"
python-multipart = "^0.0.6"
celery = {extras = ["redis"], version = "^5.3.6"}
motor = "^3.3.2"
beanie = "^1.25.0"
boto3 = "^1.34.44"
fastapi-storages = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
